move mesh tests in MRTest
